TextToSpeech/python/requirements.txt
Flask==1.1.2
pdf2image==1.13.1
Werkzeug==1.0.1
pytesseract==0.3.4
Pillow==7.2.0